perf(tests): cut assertion and setup overhead from the vitest gate
Two hotspots found by profiling the per-file timing artifacts of run
30374062348, where 20 files accounted for 50% of the suite's time.

search-console-compat was the single largest file at 47.7s. Its loops called
expect() once per path over an accumulator that has grown to 1,775,351 entries
(the assertion floor still reads 150000). Measured locally: the resolver itself
costs 8.25s over the full set, while the same loop wrapped in expect() costs
~21.8s — 13.6s is matcher construction, not work under test. The three loops now
collect offending paths and assert once. Coverage is identical, every path is
still evaluated with no sampling, and diagnostics improve: a failure reports up
to 20 offending paths plus the total instead of aborting on the first. The file
goes from 24.97s to 10.62s locally with all 30 tests green.

setupFiles runs once per test file, and tests/setup.tsx imported
@testing-library/jest-dom/vitest unconditionally. On an empty test that import
alone is the difference between 93ms and 45ms of setup — spent registering DOM
matchers for the ~1240 .test.ts files of the node project, which have no DOM to
use them on. The body moves to setup-common.tsx, shared verbatim by both
entrypoints, so nothing is duplicated; only the dom project keeps the jest-dom
import. Checked first: of those 1240 files exactly 2 mention a jest-dom matcher,
both under tests/e2e/, and both import expect from playwright/test.

Measured on the full dataset-independent group (765 files, 38299 tests, zero
failures): 74.06s to 59.52s, with setup alone dropping from 57.10s to 28.00s.
